implementing comment time
npm install dayjs
wanted to implement comment time(time when was the comment created (Time from now)) and thought about taking in moment.js library, but i think day.js would be a bit smarter to do because it seems to be a minimized version of moment.js. If i need to use only 1% of the moment.js library, then the other 99% could slow down the UI maybe?